+++ components/common/form/Form.module.css
@@ -1,0 +1,68 @@
+.form {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 2.4rem;
+}
+
+.inputName {
+  color: var(--color-black);
+  font-family: Pretendard;
+  font-size: 1.4rem;
+  font-weight: 400;
+}
+
+.inputGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 1.2rem;
+}
+
+.emailInput {
+  width: 40rem;
+  padding: 1.8rem 1.5rem;
+  justify-content: center;
+  align-items: center;
+  border-radius: 8px;
+  border: 1px solid var(--color-gray-400);
+  font-size: 1.6rem;
+}
+
+.pwGroup {
+  position: relative;
+}
+
+.pwInput {
+  width: 40rem;
+  padding: 1.8rem 1.5rem;
+  justify-content: center;
+  align-items: center;
+  border-radius: 8px;
+  border: 1px solid var(--color-gray-400);
+  font-size: 1.6rem;
+}
+
+.eye {
+  position: absolute;
+  bottom: 2.3rem;
+  right: 1.5rem;
+}
+
+.loginBtn {
+  display: flex;
+  width: 40rem;
+  padding: 1.6rem 2rem;
+  justify-content: center;
+  align-items: center;
+  gap: 1rem;
+  border-radius: 8px;
+  background: var(
+    --gra-purpleblue-to-skyblue,
+    linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%)
+  );
+  color: var(--color-gray-100);
+  font-family: Pretendard;
+  font-size: 1.8rem;
+  font-weight: 600;
+  margin-top: 0.6rem;
+}
