+++ styles/reset.css
@@ -1,7 +1,7 @@
-@import url("https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@300;400;500;600;700&display=swap");
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@300;400;500;600;700&display=swap');
 
 html {
-  font-size: var(--base-size);
+  font-size: 62.5%;
 }
 
 body {
@@ -16,7 +16,7 @@ body {
 *,
 *::before,
 *::after {
-  font-family: "Noto Sans KR", sans-serif;
+  font-family: 'Noto Sans KR', sans-serif;
   box-sizing: border-box;
   margin: 0;
   padding: 0;
@@ -55,7 +55,7 @@ button {
   -webkit-appearance: none;
 }
 
-input[type="number"] {
+input[type='number'] {
   -moz-appearance: textfield;
 }
